ch5-1 and -2: move double hash from signature to message
